separate generateMoves.test.js tests into separate files
